requirements.txt
absl-py==0.12.0
appdirs==1.4.4
asgiref==3.3.4
astunparse==1.6.3
autocorrect==2.5.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Cython==0.29.21
distlib==0.3.1
Django==3.2.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
gensim==4.0.1
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
inexactsearch==1.0.2
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
language-tool-python==2.5.3
Markdown==3.3.4
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyenchant==3.2.0
pyspellchecker==0.6.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
silpa-common==0.3
six==1.15.0
smart-open==5.0.0
soundex==1.1.3
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
textblob==0.15.3
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.6
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
